*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cairo/org/eclipse/swt/internal/cairo/cairo_font_extents_t.java
@@ -15,17 +15,17 @@
  *
  * IBM
  * -  Binding to permit interfacing between Cairo and SWT
- * -  Copyright (C) 2005 IBM Corp.  All Rights Reserved.
+ * -  Copyright (C) 2005, 2006 IBM Corp.  All Rights Reserved.
  *
  * ***** END LICENSE BLOCK ***** */
 package org.eclipse.swt.internal.cairo;
 
 
 public class cairo_font_extents_t {
-	public float ascent;
-	public float descent;
-	public float height;
-	public float max_x_advance;
-	public float max_y_advance;
+	public double ascent;
+	public double descent;
+	public double height;
+	public double max_x_advance;
+	public double max_y_advance;
 	public static final int sizeof = Cairo.cairo_font_extents_t_sizeof();
 }
